Rename filter module to filter-api
