+++ filter-api/src/main/java/org/hypertrace/agent/filter/FilterRegistry.java
@@ -21,6 +21,8 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.ServiceLoader;
 import org.hypertrace.agent.core.EnvironmentConfig;
+import org.hypertrace.agent.filter.api.Filter;
+import org.hypertrace.agent.filter.spi.FilterProvider;
 
 /**
  * Provides access to the {@link Filter} implementations. The {@link Filter} implementation are
