+++ filter-api/src/main/java/org/hypertrace/agent/filter/spi/FilterProvider.java
@@ -14,16 +14,22 @@
  * limitations under the License.
  */
 
-package org.hypertrace.agent.filter;
+package org.hypertrace.agent.filter.spi;
 
-public class ExecutionNotBlocked implements FilterResult {
+import org.hypertrace.agent.filter.api.Filter;
 
-  public static final ExecutionNotBlocked INSTANCE = new ExecutionNotBlocked();
+/**
+ * {@link FilterProvider} creates {@link Filter}.
+ *
+ * <p>The implementation is discovered via Java service loader API - each implementation has to be
+ * registered in {@code META-INF/services/}.
+ */
+public interface FilterProvider {
 
-  private ExecutionNotBlocked() {}
-
-  @Override
-  public boolean blockExecution() {
-    return false;
-  }
+  /**
+   * Create filter instance.
+   *
+   * @return a filter instance.
+   */
+  Filter create();
 }
