+++ filter-custom-opa/src/main/java/org/hypertrace/agent/filter/opa/custom/CustomOpaLib.java
@@ -25,10 +25,10 @@ import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
-import org.hypertrace.agent.filter.ExecutionBlocked;
-import org.hypertrace.agent.filter.ExecutionNotBlocked;
-import org.hypertrace.agent.filter.Filter;
-import org.hypertrace.agent.filter.FilterResult;
+import org.hypertrace.agent.filter.api.ExecutionBlocked;
+import org.hypertrace.agent.filter.api.ExecutionNotBlocked;
+import org.hypertrace.agent.filter.api.Filter;
+import org.hypertrace.agent.filter.api.FilterResult;
 import org.hypertrace.agent.filter.opa.custom.evaluator.ICustomPolicyEvaluator;
 import org.hypertrace.agent.filter.opa.custom.evaluator.IpAddressPolicyEvaluator;
 import org.slf4j.Logger;
