+++ instrumentation/grpc-1.5/src/main/java/io/opentelemetry/instrumentation/hypertrace/grpc/v1_5/server/GrpcServerInterceptor.java
@@ -34,7 +34,7 @@ import java.util.Map;
 import org.hypertrace.agent.core.HypertraceConfig;
 import org.hypertrace.agent.core.HypertraceSemanticAttributes;
 import org.hypertrace.agent.filter.FilterRegistry;
-import org.hypertrace.agent.filter.FilterResult;
+import org.hypertrace.agent.filter.api.FilterResult;
 
 public class GrpcServerInterceptor implements ServerInterceptor {
