+++ instrumentation/servlet/servlet-2.3/src/main/java/io/opentelemetry/instrumentation/hypertrace/servlet/v2_3/Servlet2BodyInstrumentation.java
@@ -44,7 +44,7 @@ import net.bytebuddy.matcher.ElementMatcher;
 import org.hypertrace.agent.core.HypertraceConfig;
 import org.hypertrace.agent.core.HypertraceSemanticAttributes;
 import org.hypertrace.agent.filter.FilterRegistry;
-import org.hypertrace.agent.filter.FilterResult;
+import org.hypertrace.agent.filter.api.FilterResult;
 
 /**
  * Body capture for servlet 2.3. Note that OTEL servlet instrumentation is compatible with servlet
