+++ javaagent-bootstrap/build.gradle.kts
@@ -8,5 +8,5 @@ dependencies{
     api("io.opentelemetry.javaagent:opentelemetry-javaagent-bootstrap:0.9.0")
     api("io.opentelemetry.javaagent:opentelemetry-javaagent-api:0.9.0")
     implementation(project(":javaagent-core"))
-    implementation(project(":filter"))
+    implementation(project(":filter-api"))
 }
